Move benchmark class.
